+++ src/main/java/org/jolokia/docker/maven/access/DockerAccess.java
@@ -2,7 +2,6 @@ package org.jolokia.docker.maven.access;
 
 import java.io.File;
 import java.util.List;
-import java.util.Map;
 
 import org.jolokia.docker.maven.access.log.LogCallback;
 import org.jolokia.docker.maven.access.log.LogGetHandle;
@@ -77,17 +76,6 @@ public interface DockerAccess {
      * @throws DockerAccessException if the container could not be stopped.
      */
     void stopContainer(String containerId) throws DockerAccessException;
-
-    /**
-     * Query the port mappings for a certain container.
-     *
-     * @param containerId id of the container to query.
-     * @return mapped ports where the keys of this map are the container ports including the protocol (e.g. "8080/tcp")
-     *         and the values are the mapped, potentially dynamically chosen, host ports (e.g. 49000).
-     *         The returned map is never null but can be empty.
-     * @throws DockerAccessException if the query failed.
-     */
-    Map<String, Integer> queryContainerPortMapping(String containerId) throws DockerAccessException;
 
     /**
      * Get logs for a container up to now synchronously.