+++ src/main/java/org/jolokia/docker/maven/model/Container.java
@@ -1,4 +1,5 @@
-package org.jolokia.docker.maven.model;/*
+package org.jolokia.docker.maven.model;
+/*
  * 
  * Copyright 2014 Roland Huss
  *
@@ -15,18 +16,42 @@ package org.jolokia.docker.maven.model;/*
  * limitations under the License.
  */
 
+import java.util.Map;
+
 /**
  * @author roland
  * @since 16/07/15
  */
 public interface Container {
 
-    String getName();
 
     long getCreated();
 
     String getId();
 
     String getImage();
+
+    String getName();
+
+    Map<String, PortBinding> getPortBindings();
+
     boolean isRunning();
+
+    class PortBinding {
+        private final String hostIp;
+        private final Integer hostPort;
+
+        public PortBinding(Integer hostPort, String hostIp) {
+            this.hostPort = hostPort;
+            this.hostIp = hostIp;
+        }
+
+        public String getHostIp() {
+            return hostIp;
+        }
+
+        public Integer getHostPort() {
+            return hostPort;
+        }
+    }
 }