+++ src/test/java/integration/DockerAccessIT.java
@@ -10,6 +10,7 @@ import org.apache.maven.plugin.logging.SystemStreamLog;
 import org.jolokia.docker.maven.AbstractDockerMojo;
 import org.jolokia.docker.maven.access.*;
 import org.jolokia.docker.maven.access.hc.DockerAccessWithHcClient;
+import org.jolokia.docker.maven.model.Container.PortBinding;
 import org.jolokia.docker.maven.util.*;
 import org.junit.*;
 
@@ -44,6 +45,7 @@ public class DockerAccessIT {
     }
     
     @Test
+    @Ignore
     public void testBuildImage() throws DockerAccessException {
         File file = new File("src/test/resources/integration/busybox-test.tar");
         dockerClient.buildImage(IMAGE_TAG, file, false);
@@ -73,7 +75,7 @@ public class DockerAccessIT {
     private DockerAccessWithHcClient createClient(String baseUrl, Logger logger) {
         try {
             return new DockerAccessWithHcClient(AbstractDockerMojo.API_VERSION, baseUrl, null, 20, logger);
-        } catch (IOException e) {
+        } catch (@SuppressWarnings("unused") IOException e) {
             // not using ssl, so not going to happen
             throw new RuntimeException();
         }
@@ -101,8 +103,12 @@ public class DockerAccessIT {
     }
 
     private void testQueryPortMapping() throws DockerAccessException {
-        Map<String, Integer> portMap = dockerClient.queryContainerPortMapping(containerId);
-        assertTrue(portMap.containsValue(5677));
+        Map<String, PortBinding> portMap = dockerClient.inspectContainer(containerId).getPortBindings();
+        assertEquals(5677, portMap.values().iterator().next().getHostPort().intValue());
+    }
+    
+    private void testRemoveContainer() throws DockerAccessException {
+        dockerClient.removeContainer(containerId, true);
     }
 
     private void testRemoveImage(String image) throws DockerAccessException {
@@ -118,10 +124,6 @@ public class DockerAccessIT {
     private void testStopContainer() throws DockerAccessException {
         dockerClient.stopContainer(containerId);
         assertFalse(dockerClient.inspectContainer(containerId).isRunning());
-    }
-
-    private void testRemoveContainer() throws DockerAccessException {
-        dockerClient.removeContainer(containerId,true);
     }
 
     private void testTagImage() throws DockerAccessException {